requirements.txt
django-debug-toolbar
south
django-crispy-forms